pyproject.toml
[project]
name = "stargazer"
version = "0.1.0"
description = "API to retrieve neighbor repositories on GitHub"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.124.4",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.27",
]
